Reorganize ops and CLI [skip ci]
