+++ microhapulator/cli/prob.py
@@ -10,6 +10,12 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
+import json
+from microhapulator import load_marker_frequencies
+from microhapulator import open as mhopen
+from microhapulator.op.prob import prob
+from microhapulator.profile import Profile
+
 
 def subparser(subparsers):
     desc = (
@@ -44,3 +50,16 @@ def subparser(subparsers):
     )
     cli.add_argument("profile1", help="profile in JSON format")
     cli.add_argument("profile2", nargs="?", default=None, help="profile in JSON format; optional")
+
+
+def main(args):
+    prof1 = Profile(fromfile=args.profile1)
+    prof2 = Profile(fromfile=args.profile2) if args.profile2 else None
+    frequencies = load_marker_frequencies(args.freq)
+    result = prob(frequencies, prof1, prof2=prof2, erate=args.erate)
+    key = "random_match_probability" if prof2 is None else "likelihood_ratio"
+    data = {
+        key: "{:.3E}".format(result),
+    }
+    with mhopen(args.out, "w") as fh:
+        json.dump(data, fh, indent=4)
