+++ microhapulator/op/balance.py
@@ -43,10 +43,3 @@ def balance(profile, include_discarded=True):
         data.to_csv(tfile, index=False, header=False)
         subprocess.run(["termgraph", tfile])
     return data
-
-
-def main(args):
-    profile = microhapulator.profile.ObservedProfile(fromfile=args.input)
-    data = balance(profile, include_discarded=args.discarded)
-    if args.csv:
-        data.to_csv(args.csv, index=False)
