+++ microhapulator/op/contain.py
@@ -10,17 +10,14 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
-import numpy.random
-import microhapulator
-from microhapulator.profile import Profile
 
-
-def main(args):
-    if args.seed:
-        numpy.random.seed(args.seed)
-    profile = Profile.unite(
-        Profile(fromfile=args.mom),
-        Profile(fromfile=args.dad),
-    )
-    with microhapulator.open(args.out, "w") as fh:
-        profile.dump(fh)
+def contain(p1, p2):
+    """Compute the proportion of alleles from p2 present in p1."""
+    total = 0
+    contained = 0
+    for marker in p2.markers():
+        allele1 = p1.alleles(marker)
+        allele2 = p2.alleles(marker)
+        total += len(allele2)
+        contained += len(allele2 & allele1)
+    return contained, total
