+++ microhapulator/op/contrib.py
@@ -10,21 +10,7 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
-import json
 from math import ceil
-import microhapulator
-from microhapulator.profile import Profile
-
-
-def load_profile(bamfile=None, markertsv=None, json=None, **kwargs):
-    if not json and (not bamfile or not markertsv):
-        message = "must provide either JSON profile or BAM and refr FASTA"
-        raise ValueError(message)
-    if json:
-        profile = Profile(fromfile=json)
-    else:
-        profile = microhapulator.type.type(bamfile, markertsv, **kwargs)
-    return profile
 
 
 def contrib(profile):
@@ -34,21 +20,3 @@ def contrib(profile):
     max_loci = sum([1 for n in num_alleles_per_marker if n >= max_thresh])
     max_perc = round(max_loci / len(num_alleles_per_marker), 4)
     return ceil(max_num_alleles / 2), max_loci, max_perc
-
-
-def main(args):
-    profile = load_profile(
-        bamfile=args.bam,
-        markertsv=args.tsv,
-        json=args.json,
-        static=args.static,
-        dynamic=args.dynamic,
-    )
-    ncontrib, nloci, ploci = contrib(profile)
-    data = {
-        "min_num_contrib": ncontrib,
-        "num_loci_max_alleles": nloci,
-        "perc_loci_max_alleles": ploci,
-    }
-    with microhapulator.open(args.out, "w") as fh:
-        json.dump(data, fh, indent=4)
