+++ microhapulator/op/diff.py
@@ -10,10 +10,6 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
-import json
-import microhapulator
-from microhapulator.profile import Profile
-
 
 def diff(prof1, prof2):
     allmarkers = set(prof1.markers()).union(prof2.markers())
@@ -24,16 +20,3 @@ def diff(prof1, prof2):
         diff2 = allele2 - allele1
         if len(diff1) > 0 or len(diff2) > 0:
             yield marker, diff1, diff2
-
-
-def main(args):
-    differ = diff(Profile(fromfile=args.profile1), Profile(fromfile=args.profile2))
-    with microhapulator.open(args.out, "w") as fh:
-        for marker, diff1, diff2 in differ:
-            print(marker, file=fh)
-            if len(diff1) > 0:
-                for haplotype in sorted(diff1):
-                    print(">>>", haplotype, file=fh)
-            if len(diff2) > 0:
-                for haplotype in sorted(diff2):
-                    print("<<<", haplotype, file=fh)
