+++ microhapulator/op/dist.py
@@ -10,29 +10,12 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
-import json
-import microhapulator
-from microhapulator.profile import Profile
 
-
-def contain(p1, p2):
-    """Compute the proportion of alleles from p2 present in p1."""
-    total = 0
-    contained = 0
-    for marker in p2.markers():
+def dist(p1, p2):
+    hammdist = 0
+    for marker in set(p1.markers()).union(p2.markers()):
         allele1 = p1.alleles(marker)
         allele2 = p2.alleles(marker)
-        total += len(allele2)
-        contained += len(allele2 & allele1)
-    return contained, total
-
-
-def main(args):
-    contained, total = contain(Profile(fromfile=args.profile1), Profile(fromfile=args.profile2))
-    data = {
-        "containment": round(contained / total, 4),
-        "contained_alleles": contained,
-        "total_alleles": total,
-    }
-    with microhapulator.open(args.out, "w") as fh:
-        json.dump(data, fh, indent=4)
+        if allele1 != allele2:
+            hammdist += 1
+    return hammdist
