+++ microhapulator/op/sim.py
@@ -36,35 +36,3 @@ def sim(frequencies, seed=None):
     message = f"simulated microhaplotype variation at {len(markers)} markers"
     microhapulator.plog("[MicroHapulator::sim]", message)
     return profile
-
-
-def load_inputs(freqfile, markerfile, seqfile, haploseqs=False):
-    frequencies = microhapulator.load_marker_frequencies(freqfile)
-    if not haploseqs:
-        return frequencies, None, None
-    markers = microhapulator.load_marker_definitions(markerfile)
-    sequences = microhapulator.load_marker_reference_sequences(seqfile)
-    microhapulator.cross_check_marker_ids(
-        frequencies.Marker, markers.Marker, "marker frequencies", "marker definitions"
-    )
-    microhapulator.cross_check_marker_ids(
-        frequencies.Marker, sequences.keys(), "marker frequencies", "marker reference sequences"
-    )
-    return frequencies, markers, sequences
-
-
-def main(args):
-    frequencies, markers, sequences = load_inputs(
-        args.freq, args.markers, args.sequences, haploseqs=args.haplo_seq
-    )
-    profile = sim(frequencies, seed=args.seed)
-    with microhapulator.open(args.out, "w") as fh:
-        profile.dump(fh)
-        message = "profile JSON written to {:s}".format(fh.name)
-        microhapulator.plog("[MicroHapulator::sim]", message)
-    if args.haplo_seq:
-        with microhapulator.open(args.haplo_seq, "w") as fh:
-            for defline, sequence in profile.haploseqs(markers, sequences):
-                print(">", defline, "\n", sequence, sep="", file=fh)
-            message = "haplotype sequences written to {:s}".format(fh.name)
-            microhapulator.plog("[MicroHapulator::sim]", message)
